Move AI Client related stuff into its own folder
