+++ app/src/lib/backend/aiService.test.ts
@@ -1,9 +1,9 @@
 import { CloudClient } from './cloud';
 import { AnthropicModelName, ModelKind, OpenAIModelName } from './types';
+import { AnthropicAIClient } from '$lib/backend/aiClient/anthropic';
+import { ButlerAIClient } from '$lib/backend/aiClient/butler';
+import { OpenAIClient } from '$lib/backend/aiClient/openAI';
 import { AIService, GitAIConfigKey, KeyOption, buildDiff } from '$lib/backend/aiService';
-import { AnthropicAIClient } from '$lib/backend/anthropic';
-import { ButlerAIClient } from '$lib/backend/butler';
-import { OpenAIClient } from '$lib/backend/openAI';
 import * as toasts from '$lib/utils/toasts';
 import { Hunk } from '$lib/vbranches/types';
 import { plainToInstance } from 'class-transformer';
