+++ app/src/lib/backend/aiService.ts
@@ -1,7 +1,7 @@
 import { AnthropicModelName, OpenAIModelName } from './types';
-import { AnthropicAIClient } from '$lib/backend/anthropic';
-import { ButlerAIClient } from '$lib/backend/butler';
-import { OpenAIClient } from '$lib/backend/openAI';
+import { AnthropicAIClient } from '$lib/backend/aiClient/anthropic';
+import { ButlerAIClient } from '$lib/backend/aiClient/butler';
+import { OpenAIClient } from '$lib/backend/aiClient/openAI';
 import { splitMessage } from '$lib/utils/commitMessage';
 import * as toasts from '$lib/utils/toasts';
 import OpenAI from 'openai';
